booth: all over tidy up ans structure everything Fixes #12
