+++ booth/tests/_fixtures/run_postprocess.py
@@ -1,4 +1,4 @@
-from src.postprocessing.postprocess import postprocess
+from src.postprocessing._postprocess import postprocess
 import os
 import unittest
